Added Policy to delete note

+++ resources/views/livewire/notes/show-notes.blade.php
@@ -7,7 +7,7 @@ new class extends Component {
     public function delete($noteId)
     {
         $note = Note::where('id', $noteId)->first();
-       // $this->authorize('delete', $note);
+        $this->authorize('delete', $note);
         $note->delete();
     }
